Replace all the files Organizations by Companies
