+++ src/api/dev/companies.http
@@ -1,0 +1,9 @@
+http://localhost:5000/api/v1/companies
+
+###
+http://localhost:5000/api/v1/companies/1
+
+###
+
+
+
