--- src/api/dev/organizations.http
@@ -1,9 +1,0 @@
-http://localhost:5000/api/v1/organizations
-
-###
-http://localhost:5000/api/v1/organizations/1
-
-###
-http://localhost:5000/api/v1/organizations/LYNXTER
-
-
