+++ src/api/dev/usersRequest.http
@@ -13,5 +13,5 @@ content-type: application/json
     "email": "formateur@test.fr",
     "role": "SUPER_ADMIN",
     "profession": "",
-    "organization_id": null
+    "company_id": null
 }
